pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "pacman-rldp"
version = "0.1.0"
description = "Tabular RL agent for a simplified Pacman environment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=2.4",
  "matplotlib>=3.8",
  "pyyaml>=6.0",
  "gymnasium>=1.0",
  "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "pre-commit>=3.7",
  "pytest>=8.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pacman_rldp"]

[tool.hatch.build.targets.sdist]
include = [
  "src/pacman_rldp",
  "README.md",
  "CONTRIBUTING.md",
  "configs",
  "scripts",
]

[tool.ruff]
line-length = 100
src = ["src", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
ignore = ["E501"]
